requirements.txt
langchain>=0.1.14
langchain-community>=0.0.21
langchain-chroma>=0.0.7
langchain-google-genai>=0.0.7
langchain-openai>=0.1.4
langchain-groq>=0.0.2
langchain-tavily>=0.0.3
langgraph>=0.0.36
chromadb>=0.4.24
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-dotenv>=1.0.1
pydantic>=2.6.4
typing-extensions>=4.10.0
google-auth>=2.29.0
groq>=0.5.0
streamlit>=1.35.0

